Move HighlightAttribute to ShellWidget
This move is necessary for the upcoming change to replace individual
properties in Cell with HighlightAttribute.
