+++ src/gui/shellwidget/highlight.cpp
